+++ Syncly/src/pages/LoginPage.tsx
@@ -7,7 +7,7 @@ import { useMutation } from "@tanstack/react-query";
 import { PostLogin } from "../shared/api/Auth";
 import { useAuthContext } from "../context/AuthContext";
 import { AxiosError } from "axios";
-import { BeginGoogleLogin } from "../shared/api/Social";
+import { Social } from "../shared/api/Social";
 import { PostPersonalSpace } from "../shared/api/WorkSpace/post";
 
 const LoginPage = () => {
@@ -131,7 +131,7 @@ const LoginPage = () => {
         {/* Google Sign-In */}
 
         <button 
-          onClick={() => BeginGoogleLogin()}
+          onClick={() => Social()}
           className="w-[459px] flex items-center justify-center gap-4 border border-[#E6E6E6] mt-4 px-4 py-2 rounded-[8px] bg-white cursor-pointer gap-2 text-black text-sm font-medium leading-6 font-['inter']">
           <img src="/google-logo.png" className="w-6 h-6" alt="Google" />{" "}
           <p>Continue with Google</p>
